move change password components to profile
